requirements.txt
google-genai
keyring
questionary
requests
rich
SecretStorage
dbus-python; sys_platform == 'linux'
